move benchmark functions with similar setup.py to a different file
