+++ test/benchmarks/pulse/schedule_construction.py
@@ -59,14 +59,6 @@ class ScheduleConstructionBench:
         self.parametric_sched = build_parametric_pulse_schedule(unique_pulses,
                                                                 channels)
 
-        qr = QuantumRegister(1)
-        self.qc = QuantumCircuit(qr)
-        self.qc.append(Gate('my_pulse', 1, []), qargs=[qr[0]])
-        self.backend = FakeOpenPulse2Q()
-        self.inst_map = self.backend.defaults().instruction_schedule_map
-        self.add_inst_map = self.inst_map
-        self.add_inst_map.add('my_pulse', [0], self.parametric_sched)
-
     def time_sample_pulse_schedule_construction(self,
                                                 unique_pulses,
                                                 channels):
@@ -84,13 +76,3 @@ class ScheduleConstructionBench:
         sched = self.sample_sched.shift(self.parametric_sched.stop_time)
         sched.insert(self.parametric_sched.start_time,
                      self.parametric_sched)
-
-    def time_build_instruction(self, _, __):
-        self.inst_map.add('my_pulse', [0], self.parametric_sched)
-
-    def time_instruction_to_schedule(self, _, __):
-        schedule(self.qc, self.backend, inst_map=self.add_inst_map)
-
-    def time_union_of_schedules(self, _, __):
-        sched = Schedule()
-        sched.union(self.sample_sched)
